refactor(pkg,agent): prepare agent to receive gateway mode
This commit focuses on restructuring the 'pkg/agent' module to pave the way for
the upcoming gateway mode feature. It is worth to say, no logic was altered
during this refactor; rather, the code was reorganized to accommodate the new
requirements.

Key changes introduced in this refactor are:
- Introduction of a new package, 'mode', which contains interfaces for
  authentication methods and session handlers.
- Creation of a 'host' package responsible for housing the implementation
  details of authentication and session handling within the 'mode' package
  when the agent is running in 'host' mode.
