+++ pkg/agent/server/modes/host/host.go
@@ -1,0 +1,2 @@
+// Package host defines authentication and sessions handles when SSH's server is running in host mode.
+package host
