+++ pkg/agent/server/modes/host/pty.go
@@ -1,4 +1,4 @@
-package server
+package host
 
 import (
 	"io"
