+++ pkg/agent/server/subsystem.go
@@ -1,0 +1,12 @@
+package server
+
+import (
+	gliderssh "github.com/gliderlabs/ssh"
+)
+
+// sftpSubsystemHandler handles the SFTP subsystem session.
+func (s *Server) sftpSubsystemHandler(session gliderssh.Session) {
+	go s.startKeepAliveLoop(session)
+
+	s.sessioner.SFTP(session) //nolint:errcheck
+}
